refactor: Moved TDPDataMatrix and CIDataMatrix into separate files to alleviate the DataContainers.groovy file
