+++ plugins/nf-co2footprint/src/main/nextflow/co2footprint/CIDataMatrix.groovy
@@ -1,0 +1,10 @@
+package nextflow.co2footprint
+
+import nextflow.co2footprint.utils.DataMatrix
+
+
+// TODO: CI Data Matrix
+class CIDataMatrix extends DataMatrix {
+
+
+}
